+++ http_testing/cookie.py
@@ -1,0 +1,3 @@
+from .assertion_elements.cookies_assertion import Cookie
+
+__all__ = ["Cookie"]