+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # Changelog
+
+## 0.3.0
+- feat: pass kwargs from PageChecker call to httpx request
+- feat: add property `previous_response` on PageChecker
+- chore: update dependencies
+
 ## 0.2.0
 - Add validators: Regex and Text
 - Rename Cookie.value_pattern to Cookie.value